Was able to grab the top songs for a user
Able to print out the top songs for a user. Still working on actually getting the artists name itself. Able to grab all the information that is inside of the TrackObject that spotify returns to us when we request it. Only issue still being worked on is being able to also grab the artists name too.